feat: adiciona dashboard evolucao ministerio e endpoint bruno

+++ diacono/src/main/java/com/diacono/diacono/ministerio/model/dto/response/MinisterioKpisResponseDTO.java
@@ -3,6 +3,6 @@ package com.diacono.diacono.ministerio.model.dto.response;
 public record MinisterioKpisResponseDTO(
 
         long ministeriosAtivos,
-        long mediaMembrosMinisterio
+        double mediaMembrosMinisterio
 ) {
 }
